backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
streamlit
langchain
openai
python-dotenv
psycopg2-binary
langchain_community